DiffStrategy is now public
